requirements.txt
pyfiglet
dnspython
requests
python-whois
beautifulsoup4
